refactor(pipeline): adopt shared CircularBuffer, relocate to core (#3288)
Pipeline EventBus used O(n) Array.shift() eviction, duplicating the O(1)
CircularBuffer. Move CircularBuffer agents/collaboration -> core (back-compat
re-export kept) and use it in the pipeline EventBus. Behavior-preserving:
same oldest-first eviction + query order. 90 tests pass.
