+++ packages/nexus-agents/src/core/circular-buffer.test.ts
